Adjust storm data

Adjusts storm data based on user selections on date range, distance limit to a storm, etc.

Usage

adjust_storm_data(
  storm_data,
  date_range = NULL,
  event_types = NULL,
  dist_limit = NULL,
  storm = NULL
)

Arguments

storm_data

A dataset of storm data. This dataset must include certain columns given in the NOAA Storm Events datasets for which this package was created.

date_range

A character vector of length two with the start and end dates to pull data for (e.g., c("1999-10-16", "1999-10-18")).

event_types

Character vector with the types of storm events that should be kept. The default value (NULL) keeps all types of events. See the "Details" vignette for this package for more details on possible event types.

dist_limit

A numeric scalar with the distance (in kilometers) that a county must be from the storm's path to be included. The default (NULL) does not eliminate any events based on distance from a storm's path. This option should only be used when also specifying a storm with the storm parameter.

storm

A character string with the name of the storm to pull storm events data for. This string must follow the format "[storm-name]-[4-digit storm year]" (e.g., "Floyd-1999"). Currently, this functionality only works for storms included in the extended hurricane best tracks, which covers 1988 to 2015.